feat!(lee): event selectors

+++ lez/indexer/ffi/src/api/types/vectors.rs
@@ -11,7 +11,7 @@ pub type FfiBlockBody = FfiVec<FfiTransaction>;
 
 pub type FfiNonceList = FfiVec<FfiNonce>;
 
-pub type FfiInstructionDataList = FfiVec<u32>;
+pub type FfiInstructionDataList = FfiVec<u8>;
 
 pub type FfiSignaturePubKeyList = FfiVec<FfiSignaturePubKeyEntry>;
